+++ stegasm/src/assembler/utils/utils.h
@@ -20,6 +20,6 @@ namespace assembler
     );
 
     bool token_is_uint32_value(const std::string &token);
-    uint16_t token_to_uint32(const std::string &token);
+    uint32_t token_to_uint32(const std::string &token);
 
 }
